Refactor configurable table data handling
