+++ src/app/(admin)/(others-pages)/(tables)/basic-tables/page.tsx
@@ -2,7 +2,8 @@ import ComponentCard from "@/components/common/ComponentCard";
 import PageBreadcrumb from "@/components/common/PageBreadCrumb";
 import { Metadata } from "next";
 import React from "react";
-import BasicTable from "@/components/tables/BasicTable";
+import ConfigurableTable from "@/components/tables/ConfigurableTable";
+import { orderTableConfig, orders } from "@/data/dummyTableData";
 
 export const metadata: Metadata = {
     title: "Next.js Basic Table | TailAdmin - Next.js Dashboard Template",
@@ -17,7 +18,7 @@ export default function BasicTables() {
             <PageBreadcrumb pageTitle="Basic Table"/>
             <div className="space-y-6">
                 <ComponentCard title="Basic Table 1">
-                    <BasicTable/>
+                    <ConfigurableTable data={orders} config={orderTableConfig} />
                 </ComponentCard>
             </div>
         </div>
