+++ src/app/(admin)/(rgs)/games/page.tsx
@@ -2,7 +2,8 @@ import PageBreadcrumb from "@/components/common/PageBreadCrumb";
 import { Metadata } from "next";
 import React from "react";
 import ComponentCard from "@/components/common/ComponentCard";
-import BasicTable from "@/components/tables/BasicTable";
+import ConfigurableTable from "@/components/tables/ConfigurableTable";
+import { orderTableConfig, orders } from "@/data/dummyTableData";
 
 export const metadata: Metadata = {
     title: "Next.js Blank Page | TailAdmin - Next.js Dashboard Template",
@@ -15,7 +16,7 @@ export default function GamesPage() {
             <PageBreadcrumb pageTitle="All Games" />
             <div className="space-y-6">
                 <ComponentCard title="Games">
-                    <BasicTable />
+                    <ConfigurableTable data={orders} config={orderTableConfig} />
                 </ComponentCard>
             </div>
         </div>
